Some changes in  question no. 2 of Assignment No. 01

+++ Problem Solving/Jiya's Sequence.cpp
@@ -18,7 +18,7 @@ using namespace std;
 int main(){
     int t;
     cin>>t;
-    int mul = 1;
+    ll mul = 1;
     while(t--){
         int n;
         cin>>n;
